backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
pydantic==2.5.0
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
kaggle==1.5.16
